# docs/openapi.yaml
openapi: 3.0.3
info:
  title: mStream API
  version: "1"
  description: |
    REST API for the [mStream](https://github.com/IrosTheBeggar/mStream) music streaming server.

    ## Authentication

    Most endpoints require a JWT bearer token. Tokens are issued by `POST /api/v1/auth/login` and
    can be supplied via any of (checked in order):

    1. Request body field `token`
    2. Query parameter `token`
    3. HTTP header `x-access-token`
    4. Cookie `x-access-token`

    ### Public mode

    If no users exist in the database, all requests are authenticated as a virtual public user with
    full library access — no token required. This is the default for fresh installs.

    ### Special tokens

    - **Admin tokens** have `admin: true` in the payload and are required for every `/api/v1/admin/*` endpoint.
    - **Jukebox tokens** (`jukebox: true`) are short-lived and bound to an active remote-control session.
    - **Shared-playlist tokens** (`shareToken: true`) restrict file access to the tracks in one specific shared playlist.
    - **Federation invite tokens** (`invite: true`) can only call `POST /api/v1/federation/invite/exchange`.

    ## Errors

    Errors use the shape `{ "error": "<message>" }`. Status codes:

    | Code | Meaning |
    |------|---------|
    | 400  | Bad request (invalid input) |
    | 401  | Unauthenticated |
    | 403  | Forbidden (Joi validation error, missing permission) |
    | 404  | Resource not found |
    | 405  | Admin API locked |
...

...

...
servers:
...
security:
  - bearerAuth: []
  - cookieAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Pass the JWT in the `x-access-token` header. Despite the OpenAPI "bearer" label,
        the value does NOT include a `Bearer ` prefix — it's the raw token.
    cookieAuth:
      type: apiKey
      in: cookie
      name: x-access-token
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message.
      example:
        error: Authentication Error
...
  responses:
...
    AdminLocked:
      description: Admin API disabled (either by configuration or because the user is not an admin).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
paths:
...
  /api/v1/discogs/coverart:
    get:
      tags: [Discogs]
      summary: Search Discogs for album art
      parameters:
        - name: artist
          in: query
          schema: { type: string }
        - name: title
          in: query
          schema: { type: string }
        - name: album
          in: query
          schema: { type: string }
        - name: year
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: Choices.
          content:
            application/json:
              schema:
                type: object
                properties:
                  choices:
                    type: array
                    items:
                      type: object
                      properties:
                        releaseId: { type: string }
                        releaseTitle: { type: string }
                        year: { type: integer }
                        thumbB64: { type: string }
                        coverImage: { type: string }
  /api/v1/deezer/search:
    get:
      tags: [Discogs]
      summary: Search Deezer (part of the Discogs art-lookup flow)
      parameters:
        - name: q
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Deezer response.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { type: object }
  /api/v1/discogs/embed:
    post:
      tags: [Discogs]
      summary: Embed album art from Discogs / Deezer / URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [filepath]
              properties:
                filepath: { type: string }
                releaseId: { type: string }
                coverUrl: { type: string }
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                type: object
                properties:
                  aaFile: { type: string, description: "Filename of the cached art." }

  # ── Server Playback (Rust audio proxy) ───────────────────────────────────
...
  /api/v1/admin/discogs/config:
    get:
      tags: [Admin - Discogs]
      summary: Get Discogs configuration
      responses:
        "200":
          description: Config.
          content:
            application/json:
              schema:
                type: object
                properties:
                  enabled: { type: boolean }
                  allowArtUpdate: { type: boolean }
                  allowId3Edit: { type: boolean }
                  apiKey: { type: string }
                  apiSecret: { type: string, description: "Masked." }
        "405": { $ref: "#/components/responses/AdminLocked" }
    post:
      tags: [Admin - Discogs]
      summary: Update Discogs configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                enabled: { type: boolean }
                allowArtUpdate: { type: boolean }
                apiKey: { type: string }
                apiSecret: { type: string }
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean, enum: [true] }
        "405": { $ref: "#/components/responses/AdminLocked" }

  # ── Admin - ID3 tags (Velvet) ────────────────────────────────────────────
...